Separate/refactor send events and query segments unit tests
